Updated module augmentation for find-and-replace package.
