+++ packages/ckeditor5-find-and-replace/src/findandreplace.ts
@@ -103,10 +103,3 @@ export default class FindAndReplace extends Plugin {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ FindAndReplace.pluginName ]: FindAndReplace;
-	}
-}
-
