+++ packages/ckeditor5-find-and-replace/src/findcommand.ts
@@ -97,9 +97,3 @@ export default class FindCommand extends Command {
 		};
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		'find': FindCommand;
-	}
-}
