+++ packages/ckeditor5-find-and-replace/src/index.ts
@@ -8,4 +8,13 @@
  */
 
 export { default as FindAndReplace } from './findandreplace';
+export { default as FindAndReplaceEditing } from './findandreplaceediting';
+export { default as FindAndReplaceUI } from './findandreplaceui';
 export { default as FindAndReplaceUtils } from './findandreplaceutils';
+export { default as FindCommand } from './findcommand';
+export { default as FindNextCommand } from './findnextcommand';
+export { default as FindPreviousCommand } from './findpreviouscommand';
+export { default as ReplaceCommand } from './replacecommand';
+export { default as ReplaceAllCommand } from './replaceallcommand';
+
+import './augmentation';
